requirements.txt
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
dask==2023.4.1
distributed==2023.4.1
fsspec==2023.4.0
gitdb==4.0.10
GitPython==3.1.31
importlib-metadata==6.6.0
Jinja2==3.1.2
joblib==1.2.0
locket==1.0.0
MarkupSafe==2.1.2
msgpack==1.0.5
numpy==1.24.3
packaging==23.1
pandas==2.0.1
partd==1.4.0
psutil==5.9.5
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
smmap==5.0.0
sortedcontainers==2.4.0
tblib==1.7.0
threadpoolctl==3.1.0
toolz==0.12.0
tornado==6.3.1
tzdata==2023.3
urllib3==2.0.2
zict==3.0.0
zipp==3.15.0
